Move HTTP functions to the Utils module
